pyproject.toml
[tool.poetry]
name = "drftest"
version = "0.1.0"
description = ""
authors = ["Makima3 <natalia.kolchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
